+++ libs/kfs/md5.c
@@ -39,6 +39,8 @@ struct KMD5File;
 #include <klib/rc.h>
 #include <klib/text.h>
 
+#include <kns/stream.h> /* KStreamSetReadObserver */
+
 #include <os-native.h>
 #include <sysalloc.h>
 
@@ -48,6 +50,8 @@ struct KMD5File;
 #include <string.h>
 #include <endian.h>
 #include <byteswap.h>
+
+#include "../kns/stream-priv.h" /* rcStream */
 
 #define MD5_DBG_REF(msg) DBGMSG(DBG_LEGREF,DBG_FLAG(DBG_LEGREF_MD5), msg)
 
@@ -1928,7 +1932,7 @@ struct KFileMD5ReadObserver {
     bool completed; /* entire file was read */
     bool finished;  /* digest is ready */
     uint8_t digest[16];
-    uint64_t size; /* file size, 0 if unknown */
+    uint64_t size; /* file size, ~0 if unknown */
     uint64_t pos; /* last processed position in file */
     MD5State md5;
 };
@@ -1961,12 +1965,15 @@ LIB_EXPORT rc_t CC KFileMD5ReadObserverRelease(
             free((void*)self);
             return 0;
         case krefNegative:
-            return RC(rcNS, rcMgr, rcAttaching, rcRefcount, rcInvalid);
+            return RC(rcFS, rcFile, rcReleasing, rcRange, rcExcessive);
         }
     }
 
     return 0;
 }
+
+static void CC file_observer_update(
+    void* self, rc_t rc, uint64_t pos, void* vbuffer, size_t num_read);
 
 LIB_EXPORT rc_t CC KFileMD5ReadObserverGetDigest(
     const KFileMD5ReadObserver *cself, uint8_t digest[16],
@@ -1977,8 +1984,14 @@ LIB_EXPORT rc_t CC KFileMD5ReadObserverGetDigest(
     if (error != NULL)
         *error = NULL;
 
+    if (digest == NULL)
+        return RC(rcFS, rcFile, rcReading, rcParam, rcNull);
+
     if (self == NULL)
         return RC(rcFS, rcFile, rcReading, rcSelf, rcNull);
+
+    if (!self->completed && self->size == 0)
+        file_observer_update(self, 0, 0, NULL, 0);
 
     if (!self->completed) {
         if (error != NULL) {
@@ -1990,8 +2003,8 @@ LIB_EXPORT rc_t CC KFileMD5ReadObserverGetDigest(
                 rc_t rc = 0;
                 char *e = malloc(num_writ + 1);
                 if (e == NULL)
-                    return RC(rcFS, rcFile, rcUpdating, rcMemory, rcExhausted);
-                if (self->size > 0)
+                    return RC(rcFS, rcFile, rcReading, rcMemory, rcExhausted);
+                if (self->size != ~0)
                     rc = string_printf(e, num_writ + 1, &num_writ,
                         "The file was not read to the end; it was read "
                         "to byte %lu of %lu.", self->pos, self->size);
@@ -2019,7 +2032,7 @@ LIB_EXPORT rc_t CC KFileMD5ReadObserverGetDigest(
     return 0;
 }
 
-static void CC read_observer_update(
+static void CC file_observer_update(
     void *self, rc_t rc, uint64_t pos, void *vbuffer, size_t num_read)
 {
     KFileMD5ReadObserver * observer = (KFileMD5ReadObserver*)self;
@@ -2031,7 +2044,8 @@ static void CC read_observer_update(
     if (observer->completed)
         return;
 
-    if (pos > observer->pos) /* a part of file was skipped */
+    if (observer->size > 0 && pos > observer->pos)
+        /* a part of file was skipped */
         return;
 
     if (pos < observer->pos) {
@@ -2051,11 +2065,11 @@ static void CC read_observer_update(
     MD5StateAppend(&observer->md5, buffer, num_read);
     observer->pos += num_read;
 
-    if (observer->size > 0 && observer->pos >= observer->size)
+    if (observer->size != ~0 && observer->pos >= observer->size)
         observer->completed = true;
 }
 
-static rc_t CC read_observer_destroy(void *self )
+static rc_t CC file_observer_destroy(void *self )
 {   return KFileMD5ReadObserverRelease(self); }
 
 LIB_EXPORT rc_t CC KFileMakeMD5ReadObserver(const KFile *self,
@@ -2077,14 +2091,14 @@ LIB_EXPORT rc_t CC KFileMakeMD5ReadObserver(const KFile *self,
 
     rc = KFileSize(self, &obj->size);
     if (rc != 0)
-        obj->size = 0;
+        obj->size = ~0;
 
     KRefcountInit(&obj->refcount, 1, "KFileMD5ReadObserver", "make", "file");
 
     MD5StateInit(&obj->md5);
 
-    rc = KFileSetReadObserver(
-        (KFile*)self, read_observer_update, read_observer_destroy, obj);
+    rc = KFileSetReadObserver((KFile*)self,
+        file_observer_update, file_observer_destroy, obj);
 
     if (rc == 0)
         rc = KFileMD5ReadObserverAddRef(obj);
@@ -2095,4 +2109,147 @@ LIB_EXPORT rc_t CC KFileMakeMD5ReadObserver(const KFile *self,
     return rc;
 }
 
-/******************************************************************************/
+
+/*
+ * Observer to calculate md5 checksum when KStream is read
+ */
+
+struct KStreamMD5ReadObserver {
+    KRefcount refcount;
+    bool completed; /* EOF was detected */
+    bool finished;  /* digest is ready */
+    uint8_t digest[16];
+    rc_t rc; /* returned by KStreamRead */
+    MD5State md5;
+};
+
+LIB_EXPORT rc_t CC KStreamMD5ReadObserverAddRef(
+    const KStreamMD5ReadObserver *self)
+{
+    if (self != NULL) {
+        switch (KRefcountAdd(&self->refcount, "KStreamMD5ReadObserver"))
+        {
+        case krefLimit:
+            return RC(rcFS, rcStream, rcAttaching, rcRange, rcExcessive);
+        case krefNegative:
+            return RC(rcFS, rcStream, rcAttaching, rcSelf, rcInvalid);
+        default:
+            break;
+        }
+    }
+
+    return 0;
+}
+
+LIB_EXPORT rc_t CC KStreamMD5ReadObserverRelease(
+    const KStreamMD5ReadObserver* self)
+{
+    if (self != NULL) {
+        switch (KRefcountDrop(&self->refcount, "KStreamMD5ReadObserver")) {
+        case krefWhack:
+            memset((void*)self, 0, sizeof * self);
+            free((void*)self);
+            return 0;
+        case krefNegative:
+            return RC(rcFS, rcStream, rcReleasing, rcRange, rcExcessive);
+        }
+    }
+
+    return 0;
+}
+
+static void CC stream_observer_update(
+    void* self, rc_t rc, void* vbuffer, size_t num_read)
+{
+    KStreamMD5ReadObserver* observer = (KStreamMD5ReadObserver*)self;
+    unsigned char* buffer = (unsigned char*)vbuffer;
+
+    if (observer == NULL)
+        return;
+
+    if (observer->rc != 0)
+        return;
+
+    if (rc != 0) {
+        observer->rc = rc;
+        return;
+    }
+
+    if (num_read == 0)
+        observer->completed = true;
+
+    if (observer->completed)
+        return;
+
+    MD5StateAppend(&observer->md5, buffer, num_read);
+}
+
+static rc_t CC stream_observer_destroy(void* self)
+{   return KStreamMD5ReadObserverRelease(self); }
+
+LIB_EXPORT rc_t CC KStreamMD5ReadObserverGetDigest(
+    const KStreamMD5ReadObserver* cself, uint8_t digest[16], rc_t* rc)
+{
+    KStreamMD5ReadObserver* self = (KStreamMD5ReadObserver*)cself;
+
+    if (rc != NULL)
+        *rc = self->rc;
+
+    if (digest == NULL)
+        return RC(rcFS, rcStream, rcReading, rcParam, rcNull);
+
+    if (self == NULL)
+        return RC(rcFS, rcStream, rcReading, rcSelf, rcNull);
+
+    if (self->rc != 0)
+        return RC(rcFS, rcStream, rcReading, rcError, rcExists);
+
+    if (!self->completed)
+        return RC(rcFS, rcStream, rcReading, rcStream, rcIncomplete);
+
+    if (!self->finished)
+        stream_observer_update(self, 0, NULL, 0);
+
+    if (!self->finished) {
+        MD5StateFinish(&self->md5, self->digest);
+        self->finished = true;
+    }
+
+    memmove(digest, self->digest, sizeof self->digest);
+    return 0;
+}
+
+LIB_EXPORT rc_t CC KStreamMakeMD5ReadObserver(const KStream* self,
+    const KStreamMD5ReadObserver** observer)
+{
+    rc_t rc = 0;
+    KStreamMD5ReadObserver* obj = NULL;
+
+    if (observer == NULL)
+        return RC(rcFS, rcStream, rcUpdating, rcParam, rcNull);
+    *observer = NULL;
+
+    if (self == NULL)
+        return RC(rcFS, rcStream, rcUpdating, rcSelf, rcNull);
+
+    obj = calloc(1, sizeof * obj);
+    if (obj == NULL)
+        return RC(rcFS, rcStream, rcUpdating, rcMemory, rcExhausted);
+
+    KRefcountInit(&obj->refcount, 1,
+        "KStreamMakeMD5ReadObserver", "make", "stream");
+    MD5StateInit(&obj->md5);
+
+    rc = KStreamSetReadObserver((KStream*)self,
+        stream_observer_update, stream_observer_destroy, obj);
+
+    if (rc == 0)
+        rc = KStreamMD5ReadObserverAddRef(obj);
+
+    if (rc == 0)
+        *observer = obj;
+    else 
+        KStreamMD5ReadObserverRelease(obj);
+
+    return rc;
+}