+++ test/cipher/test.sh
@@ -33,6 +33,7 @@ fi
 
 #installing cipher module into newly created virtual env
 tmp_py_env=$(pwd)/temp_env_${OS}
+rm -f $tmp_py_env
 
 if [ "${VIRTUALENV}" != "" ]; then
    virtualenv -p ${PYTHON} $tmp_py_env || exit 1
@@ -42,11 +43,25 @@ fi
 . $tmp_py_env/bin/activate             || exit 2
 
 #now inside the virtual env, python is python
+if [ "${OS}" = "mac" ]; then
+  pip install --upgrade pip setuptools wheel
+fi
 
 # the following creates "build dist .eggs" in $CIPHER_DIR
 tmp_cur_dir=$(pwd)
 cd $CIPHER_DIR                       || exit 3
-python -m pip install --use-pep517 . || exit 4
+if [ "${OS}" = "mac" ]; then
+  PFX="python $CIPHER_DIR/"
+  python -m pip install --no-build-isolation . || exit 4
+else
+  PFX=
+  lsb_release=`lsb_release -is`
+  if [ "$lsb_release" = "Ubuntu" ]; then
+    python -m pip install              . || exit 4
+  else
+    python -m pip install --use-pep517 . || exit 4
+  fi
+fi
 cd $tmp_cur_dir                      || exit 5
 unset tmp_cur_dir
 
@@ -64,8 +79,8 @@ do
     echo "Hello world $i" >> test.in
 done
 
-encrypt.py --password=password123 test.in  test.enc || exit 7
-decrypt.py --password=password123 test.enc test.out || exit 8
+${PFX}encrypt.py --password=password123 test.in  test.enc || exit 7
+${PFX}decrypt.py --password=password123 test.enc test.out || exit 8
 
 diff test.in test.out
 exit_code=$?